#160 Refactor: rename folder helpers, define a method helper and remove redundant code
